// src/app/models/pokemons/personajes.model.ts
export interface Personaje{
    id: number;
    heroname: string;
    hero_img: string;
    name: string;
    role: string;
    description: string;
    primaryAtack: string;
    pA_img: string;
    secondaryAtack: string;
    sA_img: string;
    Ability_1: string;
    a1_img: string;
    Ability_2: string;
    a2_img: string;
    Ability_3: string;
    a3_img: string;
    UltimateAbility: string;
    a4_img: string;
  }